Adds new class child
