+++ packages/server/src/async/producers/measurable-state-changed.spec.js
@@ -1,0 +1,7 @@
+const index = require('./measurable-state-changed');
+
+describe('Measurable State Changed Producer', () => {
+  it('returns an constructor', () => {
+    expect(index).toBeInstanceOf(Object);
+  });
+});
