--- packages/server/src/async/producers/measurable-state.spec.js
@@ -1,7 +1,0 @@
-const index = require('./measurable-state');
-
-describe('Measurable State Producer', () => {
-  it('returns an constructor', () => {
-    expect(index).toBeInstanceOf(Object);
-  });
-});
